# [Mempool] optimize fullnode broadcast hops for latency