--- a/config/src/config/mempool_config.rs
+++ b/config/src/config/mempool_config.rs
@@ -60,11 +60,11 @@ impl Default for MempoolConfig {
         MempoolConfig {
             shared_mempool_tick_interval_ms: 50,
             shared_mempool_backoff_interval_ms: 30_000,
-            shared_mempool_batch_size: 200,
+            shared_mempool_batch_size: 300,
             shared_mempool_max_batch_bytes: MAX_APPLICATION_MESSAGE_SIZE as u64,
             shared_mempool_ack_timeout_ms: 2_000,
             shared_mempool_max_concurrent_inbound_syncs: 4,
-            max_broadcasts_per_peer: 2,
+            max_broadcasts_per_peer: 20,
             max_network_channel_size: 1024,
             mempool_snapshot_interval_secs: 180,
             capacity: 2_000_000,
@@ -103,16 +103,22 @@ impl ConfigOptimizer for MempoolConfig {
 
         // Change the default configs for VFNs
         let mut modified_config = false;
+        if node_type.is_validator() {
+            // Set the max_broadcasts_per_peer to 2 (default is 20)
+            if local_mempool_config_yaml["max_broadcasts_per_peer"].is_null() {
+                mempool_config.max_broadcasts_per_peer = 2;
+                modified_config = true;
+            }
+            // Set the batch size per broadcast to 200 (default is 300)
+            if local_mempool_config_yaml["shared_mempool_batch_size"].is_null() {
+                mempool_config.shared_mempool_batch_size = 200;
+                modified_config = true;
+            }
+        }
         if node_type.is_validator_fullnode() {
             // Set the shared_mempool_max_concurrent_inbound_syncs to 16 (default is 4)
             if local_mempool_config_yaml["shared_mempool_max_concurrent_inbound_syncs"].is_null() {
                 mempool_config.shared_mempool_max_concurrent_inbound_syncs = 16;
-                modified_config = true;
-            }
-
-            // Set the max_broadcasts_per_peer to 4 (default is 2)
-            if local_mempool_config_yaml["max_broadcasts_per_peer"].is_null() {
-                mempool_config.max_broadcasts_per_peer = 4;
                 modified_config = true;
             }
 
@@ -158,9 +164,9 @@ mod tests {
             mempool_config.shared_mempool_max_concurrent_inbound_syncs,
             16
         );
-        assert_eq!(mempool_config.max_broadcasts_per_peer, 4);
+        assert_eq!(mempool_config.max_broadcasts_per_peer, 20);
         assert_eq!(mempool_config.default_failovers, 0);
-        assert_eq!(mempool_config.shared_mempool_batch_size, 200);
+        assert_eq!(mempool_config.shared_mempool_batch_size, 300);
         assert_eq!(mempool_config.shared_mempool_tick_interval_ms, 10);
     }
 
@@ -177,7 +183,7 @@ mod tests {
             ChainId::mainnet(),
         )
         .unwrap();
-        assert!(!modified_config);
+        assert!(modified_config);
 
         // Verify that all relevant fields are not modified
         let mempool_config = &node_config.mempool;
@@ -186,18 +192,12 @@ mod tests {
             mempool_config.shared_mempool_max_concurrent_inbound_syncs,
             default_mempool_config.shared_mempool_max_concurrent_inbound_syncs
         );
-        assert_eq!(
-            mempool_config.max_broadcasts_per_peer,
-            default_mempool_config.max_broadcasts_per_peer
-        );
+        assert_eq!(mempool_config.max_broadcasts_per_peer, 2);
         assert_eq!(
             mempool_config.default_failovers,
             default_mempool_config.default_failovers
         );
-        assert_eq!(
-            mempool_config.shared_mempool_batch_size,
-            default_mempool_config.shared_mempool_batch_size
-        );
+        assert_eq!(mempool_config.shared_mempool_batch_size, 200);
         assert_eq!(
             mempool_config.shared_mempool_tick_interval_ms,
             default_mempool_config.shared_mempool_tick_interval_ms
@@ -207,16 +207,24 @@ mod tests {
     #[test]
     fn test_optimize_vfn_config_no_overrides() {
         // Create the default validator config
+        let local_shared_mempool_max_concurrent_inbound_syncs = 1;
+        let local_max_broadcasts_per_peer = 1;
         let mut node_config = NodeConfig::get_default_vfn_config();
+        node_config
+            .mempool
+            .shared_mempool_max_concurrent_inbound_syncs =
+            local_shared_mempool_max_concurrent_inbound_syncs;
+        node_config.mempool.max_broadcasts_per_peer = local_max_broadcasts_per_peer;
 
         // Create a local config YAML with some local overrides
-        let local_config_yaml = serde_yaml::from_str(
+        let local_config_yaml = serde_yaml::from_str(&format!(
             r#"
             mempool:
-              shared_mempool_max_concurrent_inbound_syncs: 4
-              max_broadcasts_per_peer: 1
+                shared_mempool_max_concurrent_inbound_syncs: {}
+                max_broadcasts_per_peer: {}
             "#,
-        )
+            local_shared_mempool_max_concurrent_inbound_syncs, local_max_broadcasts_per_peer
+        ))
         .unwrap();
 
         // Optimize the config and verify modifications are made
@@ -234,12 +242,11 @@ mod tests {
         let default_mempool_config = MempoolConfig::default();
         assert_eq!(
             mempool_config.shared_mempool_max_concurrent_inbound_syncs,
-            4
+            local_shared_mempool_max_concurrent_inbound_syncs
         );
-        assert_eq!(mempool_config.max_broadcasts_per_peer, 2);
-        assert_ne!(
-            mempool_config.default_failovers,
-            default_mempool_config.default_failovers
+        assert_eq!(
+            mempool_config.max_broadcasts_per_peer,
+            local_max_broadcasts_per_peer
         );
         assert_ne!(
             mempool_config.shared_mempool_tick_interval_ms,